refactor: 어드민 메뉴 코드 집중화 #6
