+++ src/features/admin/components/Layout/Sidebar.tsx
@@ -1,89 +1,14 @@
 import styled from "@emotion/styled";
-import { Menu, Layout, Space, Badge } from "antd";
-import type { MenuProps } from "antd";
-import {
-  IconoirProvider,
-  HomeSimpleDoor,
-  BubbleStar,
-  User,
-  HeadsetHelp,
-  Tv,
-} from "iconoir-react";
+import { Layout } from "antd";
+import { IconoirProvider } from "iconoir-react";
 
+import Menus from "./Menus";
 import Profile from "./Profile";
 import SidebarToggle from "./SidebarToggle";
 
 const { Sider } = Layout;
 const sidebarWith = 256; // 사이드바 넓이
 const collapsedWidth = 80; // 사이드바 줄었을때 넓이
-
-type CreateItemProps = {
-  key: string;
-  label: React.ReactNode;
-  icon?: React.ReactNode;
-  style?: React.CSSProperties;
-  children?: CreateItemProps[];
-};
-
-function createItem(
-  key: string,
-  label: React.ReactNode,
-  icon?: React.ReactNode,
-  style?: React.CSSProperties,
-  children?: CreateItemProps[],
-) {
-  return {
-    key,
-    label,
-    icon,
-    style,
-    children,
-  };
-}
-
-const items: MenuProps["items"] = [
-  createItem(
-    "1",
-    "홈",
-    <span>
-      <HomeSimpleDoor width={"1rem"} height={"1rem"} />
-    </span>,
-    { fontSize: "1rem" },
-  ),
-  {
-    type: "divider",
-  },
-  createItem("2", "애니메이션", <Tv />, { fontSize: "1rem" }, [
-    createItem("anime-1", "목록", null, { fontSize: "0.875rem" }),
-    createItem("anime-2", "장르", null, { fontSize: "0.875rem" }),
-    createItem("anime-3", "제작", null, { fontSize: "0.875rem" }),
-    createItem("anime-4", "작가", null, { fontSize: "0.875rem" }),
-  ]),
-  createItem("3", "리뷰", <BubbleStar />, { fontSize: "1rem" }, [
-    createItem("review-1", "목록", null, { fontSize: "0.875rem" }),
-  ]),
-  createItem("4", "회원", <User />, { fontSize: "1rem" }, [
-    createItem("user-1", "목록", null, { fontSize: "0.875rem" }),
-  ]),
-  createItem("5", "고객센터", <HeadsetHelp />, { fontSize: "1rem" }, [
-    createItem(
-      "helpdesk-1",
-      <Space>
-        문의 <Badge count={999} />
-      </Space>,
-      null,
-      { fontSize: "0.875rem" },
-    ),
-    createItem(
-      "helpdesk-2",
-      <Space>
-        신고 <Badge count={999} />
-      </Space>,
-      null,
-      { fontSize: "0.875rem" },
-    ),
-  ]),
-];
 
 const Header = styled.div<{ isCollapsed: boolean }>(
   {
@@ -150,12 +75,7 @@ export default function Sidebar({ isCollapsed, setIsCollapsed }: SidebarProps) {
           height: "1em",
         }}
       >
-        <Menu
-          mode="inline"
-          defaultSelectedKeys={["1"]}
-          items={items}
-          style={{ marginTop: "1.25rem" }}
-        />
+        <Menus style={{ marginTop: "1.25rem" }} />
         <SidebarToggle
           width={sidebarWith}
           collapsedWidth={collapsedWidth}
